Add tests for image argcompletion, moved test helpers to helper.py
